Fixed #20486 -- Ensure that file_move_safe raises an error if the destination already exists.
Thanks to kux for the report, and Russ Webber for the patch.

+++ django/core/files/move.py
@@ -51,6 +51,10 @@ def file_move_safe(old_file_name, new_file_name, chunk_size = 1024*64, allow_ove
         return
 
     try:
+        # If the destination file exists and allow_overwrite is False then raise an IOError
+        if not allow_overwrite and os.access(new_file_name, os.F_OK):
+            raise IOError("Destination file %s exists and allow_overwrite is False" % new_file_name)
+
         os.rename(old_file_name, new_file_name)
         return
     except OSError:
